requirements.txt
Pillow==9.5.0
Unidecode>=1.3.6
uvicorn[standard]
requests
one-piece-wanted-poster
fastapi
# https://te.legra.ph/file/14632fbfeea4766941a14.png